17 commit - confirmação de exclusão e alterações

+++ administrador/ConsultarLinhaPesquisa.php
@@ -31,9 +31,7 @@ $resultado_consultaLinhaPesquisa = mysqli_query($con, $result_consultaLinhaPesqu
 
 <div class="column content">
       
-<div class="title_left">
-                <h3>Consulta Linha de Pesquisa</h3>
-              </div>
+                <h3><center>Consulta Linha de Pesquisa</center></h3>
 
           
             </div>
@@ -42,7 +40,7 @@ $resultado_consultaLinhaPesquisa = mysqli_query($con, $result_consultaLinhaPesqu
                     <table id="datatable" class="table table-striped table-bordered">
                       <thead>
                         <tr>
-                        <th scope="col">Nome</th>
+                        <th scope="col">Linha de Pesquisa</th>
                         <th scope="col">Descrição</th>
                         <th scope="col">Curso</th>
                         <th scope="col"></th>
@@ -55,11 +53,11 @@ $resultado_consultaLinhaPesquisa = mysqli_query($con, $result_consultaLinhaPesqu
     <tr>
 		<td><?php echo $rows_consultaLinhaPesquisa['nomePesquisa'];?></td>
 		<td><?php echo $rows_consultaLinhaPesquisa['descricao'];?></td>
-		<td><?php echo utf8_encode($rows_consultaLinhaPesquisa['nomeCurso']);?></td>
+		<td><?php echo ($rows_consultaLinhaPesquisa['nomeCurso']);?></td>
        
 	<td>
      <?php  echo "<a class='btn btn-success'  href='DadosLinhaPesquisa.php?idLinhaPesquisa=" .$rows_consultaLinhaPesquisa['idLinhaPesquisa'] .  "'>Editar</a>";  ?>
-    <?php  echo "<a  class='btn btn-danger' href='ExcluirLinhaPesquisa.php?idLinhaPesquisa=" .$rows_consultaLinhaPesquisa['idLinhaPesquisa']. "'> Excluir</a>";  ?>
+    <?php  echo "<a  class='btn btn-danger' href='ExcluirLinhaPesquisa.php?idLinhaPesquisa=" .$rows_consultaLinhaPesquisa['idLinhaPesquisa']. "' onclick=\"return confirm('Tem certeza que deseja deletar esse registro?');\"> Excluir</a>";  ?>
 	</td>
 	</tr>
